26/03/2024

+++ src/controllers/customer-controllers/order.js
@@ -2,11 +2,11 @@ const queryMysql = require("../../database/mysql.js");
 class OrderController {
   async createOrder(req, res) {
     try {
+      let flag = false;
       // Insert the order into DONHANG table
       const order = await queryMysql(
         `INSERT INTO DONHANG (ND_id, DH_tongTien, DH_ngayDat, DH_trangthai, DH_phuongThucTT) VALUES (${req.body.ND_id}, ${req.body.tongtien}, CURDATE(), '${req.body.trangthai}', '${req.body.PTTT}')`
       );
-      console.log("Created order:", order);
 
       // Obtain the last inserted ID (DH_id)
       const orderId = order.insertId;
@@ -24,35 +24,37 @@ class OrderController {
           FROM SANPHAM
           WHERE SP_id = ${product.SP_id}
         `);
-
         // Extract the quantity from the result (assuming SP_soLuong is returned as a property of the result)
         const oldQuantity = oldQuantityProduct[0].SP_soLuong;
 
         // Calculate new quantity
         const currentQuantityProduct = oldQuantity - parseInt(product.soLuong);
-
+        console.log(currentQuantityProduct);
+        if (currentQuantityProduct <= 0) {
+          flag = true;
+          return res
+            .status(400)
+            .json({ error: "Product quantity is not sufficient" });
+        }
         // Update product quantity
         await queryMysql(`
           UPDATE SANPHAM
           SET SP_soLuong = ${currentQuantityProduct}
           WHERE SP_id = ${product.SP_id}
         `);
-
-        console.log(`Updated quantity for product with ID ${product.SP_id}`);
-        // console.log("Created order detail:", orderDetail);
       });
 
-      const giohang = await queryMysql(
-        `SELECT * FROM giohang WHERE ND_id = ${req.body.ND_id}`
-      );
+      if (!flag) {
+        const giohang = await queryMysql(
+          `SELECT * FROM giohang WHERE ND_id = ${req.body.ND_id}`
+        );
 
-      const giohangId = giohang[0].GH_id;
+        const giohangId = giohang[0].GH_id;
 
-      // console.log(giohang[0].ND_id);
+        await queryMysql(`DELETE FROM sp_giohang WHERE GH_id = ${giohangId}`);
 
-      await queryMysql(`DELETE FROM sp_giohang WHERE GH_id = ${giohangId}`);
-
-      res.status(200).json({ message: "Order created successfully" });
+        return res.status(200).json({ message: "Order created successfully" });
+      }
     } catch (error) {
       console.error("Error creating order:", error);
       res.status(500).json({ error: "Failed to create order" });
